chore: switch to struct embedding for posix flag parsing
This revision moves away from the getopt 'flag' package implementation
for a number of reasons, the primary reason being that the package name
'flag' was often confused for the standard lib or vice versa. The
package is now named 'getopt'. Additionally, the package now wraps the
standard flag.FlagSet instead of reimplementing it completely, which is
easier to maintain and allows you to easily switch between flag parsing
or getopt-style flag parsing at runtime.

Some of the flag Value examples were promoted from examples to a core
feature of the package (MapVar, TimeVar). A new Hidden type was
introduced for hiding flags from usage texts.
