+++ getopt/doc.go
@@ -1,0 +1,4 @@
+/*
+Package getopt offers utilities for parsing getopt-style command-line options (flags).
+*/
+package getopt
